dev: move preview code
